Move ProjectUsers out of ProjectSettings folder
