+++ src/components/ProjectSettings/ProjectArchive.tsx
@@ -1,13 +1,13 @@
 import { Button } from "@mui/material";
 import { ButtonProps } from "@mui/material/Button";
-import React, { useState } from "react";
+import { useState } from "react";
 import { useTranslation } from "react-i18next";
 
 import { archiveProject, restoreProject } from "backend";
 import { ButtonConfirmation } from "components/Dialogs";
 import { themeColors } from "types/theme";
 
-interface ProjectButtonWithConfirmationProps {
+interface ProjectArchiveProps {
   archive?: boolean;
   projectId: string;
   updateParent: () => void | Promise<void>;
@@ -17,8 +17,8 @@ interface ProjectButtonWithConfirmationProps {
 /**
  * Button for archiving/restoring project (changing isActive)
  */
-export default function ProjectButtonWithConfirmation(
-  props: ButtonProps & ProjectButtonWithConfirmationProps
+export default function ProjectArchive(
+  props: ButtonProps & ProjectArchiveProps
 ) {
   const [open, setOpen] = useState(false);
   const { t } = useTranslation();
@@ -41,7 +41,7 @@ export default function ProjectButtonWithConfirmation(
   }
 
   return (
-    <React.Fragment>
+    <>
       <Button
         variant="contained"
         color={props.warn ? "secondary" : "primary"}
@@ -68,6 +68,6 @@ export default function ProjectButtonWithConfirmation(
           props.archive ? "archive" : "restore"
         }-confirm`}
       />
-    </React.Fragment>
+    </>
   );
 }
