+++ src/components/ProjectUsers/AddProjectUsers.tsx
@@ -7,8 +7,8 @@ import { toast } from "react-toastify";
 import { Role, User } from "api/models";
 import * as backend from "backend";
 import { asyncRefreshProjectUsers } from "components/Project/ProjectActions";
-import EmailInvite from "components/ProjectSettings/ProjectUsers/EmailInvite";
-import UserList from "components/ProjectSettings/ProjectUsers/UserList";
+import EmailInvite from "components/ProjectUsers/EmailInvite";
+import UserList from "components/ProjectUsers/UserList";
 import { StoreState } from "types";
 import { useAppDispatch, useAppSelector } from "types/hooks";
 import { RuntimeConfig } from "types/runtimeConfig";
