+++ src/components/ProjectUsers/tests/EmailInvite.test.tsx
@@ -3,7 +3,7 @@ import renderer from "react-test-renderer";
 import "tests/reactI18nextMock.ts";
 
 import { LoadingDoneButton } from "components/Buttons";
-import EmailInvite from "components/ProjectSettings/ProjectUsers/EmailInvite";
+import EmailInvite from "components/ProjectUsers/EmailInvite";
 
 jest.mock("backend", () => ({
   emailInviteToProject: () => mockEmailInviteToProject(),
