+++ src/components/ProjectUsers/tests/UserList.test.tsx
@@ -4,7 +4,7 @@ import renderer from "react-test-renderer";
 import "tests/reactI18nextMock.ts";
 
 import { User } from "api/models";
-import UserList from "components/ProjectSettings/ProjectUsers/UserList";
+import UserList from "components/ProjectUsers/UserList";
 import { newUser } from "types/user";
 
 jest.mock("backend", () => ({
