+++ src/types/user.ts
@@ -16,3 +16,12 @@ export function newUser(name = "", username = "", password = ""): User {
     isAdmin: false,
   };
 }
+
+export function doesTextMatchUser(text: string, user: User): boolean {
+  const lower = text.toLocaleLowerCase();
+  return (
+    user.name.toLocaleLowerCase().includes(lower) ||
+    user.username.toLocaleLowerCase().includes(lower) ||
+    user.email.toLocaleLowerCase().includes(lower)
+  );
+}
